add initial implementation using spacy, refactor package structure
