+++ dialog-acts/src/tests/dialog_acts/regex/test_negations.py
@@ -1,6 +1,6 @@
 import pytest
 
-from dialog_acts.regex.negations import RegexNegationDetector
+from textability.dialog_acts.regex.negations import RegexNegationDetector
 
 class TestRegexNegationDetector:
