+++ dialog-acts/src/textability/dialog_acts/regex/negations.py
@@ -1,10 +1,9 @@
 import re
 
-from dialog_acts.base_detectors import DialogActDetector
-from dialog_acts.regex.lexicon.language_resources import (
-    YES_WORDS_NEGATABLE_LHS, YES_WORDS_NEGATABLE_RHS,
-    NO_WORDS_NON_NEGATABLE)
-from dialog_acts.regex.util import find_negation_word
+from textability.dialog_acts.base_detectors import DialogActDetector
+from textability.dialog_acts.regex.lexicon.language_resources import (
+    YES_WORDS_NEGATABLE_LHS, YES_WORDS_NEGATABLE_RHS, NO_WORDS_NON_NEGATABLE)
+from textability.dialog_acts.regex.util import find_negation_word
 
 class RegexNegationDetector(DialogActDetector):
